collector/requirements.txt
pyyaml>=6.0.1
requests>=2.32.0
psycopg[binary]>=3.2.1
